Polish the Minimal App a bit more
Not sure how much I like this commit as I may have made it a bit less exciting. I feel like this part needs to start with more of a bang, to get web devs excited, and capture some of the initial simplicity and joy that @pydanny had in his version originally.