Edits to UI instructions and on Stripe concepts
This edit corrects the instructions for how to interact with the
Stripe UI, to make them a bit more foolproof.

It also clarifies some key concepts, such as the distinction between
the the two secrets, Stripe API key and the webhook signing secret,
and the relationship between the webhook route, and the success route.